+++ src/core/bigMunkres.ts
@@ -1,12 +1,13 @@
-import { IndexArray } from "../../types/indexArray";
-import { Matrix } from "../../types/matrix";
-import { MatrixLike } from "../../types/matrixLike";
-import { MunkresResult } from "../../types/munkres";
+import { IndexArray } from "../types/indexArray";
+import { Matrix } from "../types/matrix";
+import { MatrixLike } from "../types/matrixLike";
+import { MunkresResult } from "../types/munkresResult";
 
-import { getMin } from "../arrayLike";
-import { toString as _toString, from, transpose } from "../matrix";
+import { getMin } from "../utils/arrayLike";
+import { from, transpose } from "../utils/matrix";
+import { findUncoveredMin, step5 } from "./numMunkres";
 
-export function safeExec(matrix: MatrixLike<number>): MunkresResult<number> {
+export function safeExec(matrix: MatrixLike<bigint>): MunkresResult<bigint> {
   // Get dimensions
   const Y = matrix.length;
   const X = matrix[0]?.length ?? 0;
@@ -14,14 +15,14 @@ export function safeExec(matrix: MatrixLike<number>): MunkresResult<number> {
   // Transpose if Y > X
   if (Y > X) {
     matrix = from(matrix);
-    transpose(matrix as Matrix<number>);
+    transpose(matrix as Matrix<bigint>);
   }
 
   // Get optimal assignments
   return exec(matrix);
 }
 
-export function exec(matrix: MatrixLike<number>): MunkresResult<number> {
+export function exec(matrix: MatrixLike<bigint>): MunkresResult<bigint> {
   const Y = matrix.length;
   const X = matrix[0]?.length ?? 0;
 
@@ -36,8 +37,8 @@ export function exec(matrix: MatrixLike<number>): MunkresResult<number> {
   }
 
   // Step 1: Reduce
-  const dualX = new Array<number>(X);
-  const dualY = new Array<number>(Y);
+  const dualX = new Array<bigint>(X);
+  const dualY = new Array<bigint>(Y);
   step1(matrix, dualX, dualY);
 
   // Steps 2 & 3: Find initial matching
@@ -65,9 +66,9 @@ export function exec(matrix: MatrixLike<number>): MunkresResult<number> {
  * @param dualY - The dual variables associated with each row of the matrix. Modified in place.
  */
 export function step1(
-  matrix: MatrixLike<number>,
-  dualX: number[],
-  dualY: number[]
+  matrix: MatrixLike<bigint>,
+  dualX: bigint[],
+  dualY: bigint[]
 ): void {
   const X = dualX.length;
   const Y = dualY.length;
@@ -79,20 +80,20 @@ export function step1(
 
   // Reduce columns
   if (Y < X) {
-    dualX.fill(0);
+    dualX.fill(0n);
     return;
   }
 
   let dy = dualY[0];
   let row = matrix[0];
   for (let x = 0; x < X; ++x) {
-    dualX[x] = row[x] - dy || 0;
+    dualX[x] = row[x] - dy;
   }
   for (let y = 1; y < Y; ++y) {
     dy = dualY[y];
     row = matrix[y];
     for (let x = 0; x < X; ++x) {
-      const dx = row[x] - dy || 0;
+      const dx = row[x] - dy;
       if (dx < dualX[x]) {
         dualX[x] = dx;
       }
@@ -104,15 +105,17 @@ export function step1(
  * Finds an initial matching for the munkres algorithm.
  *
  * @param matrix - The cost matrix.
+ * @param dualX - The dual variables associated with each column of the matrix.
+ * @param dualY - The dual variables associated with each row of the matrix.
  * @param starsX - An array mapping star columns to row. Modified in place.
  * @param starsY - An array mapping star rows to columns. Modified in place.
  *
  * @returns The number of matches (stars) found.
  */
 export function steps2To3(
-  matrix: MatrixLike<number>,
-  dualX: ArrayLike<number>,
-  dualY: ArrayLike<number>,
+  matrix: MatrixLike<bigint>,
+  dualX: ArrayLike<bigint>,
+  dualY: ArrayLike<bigint>,
   starsX: number[],
   starsY: number[]
 ): number {
@@ -124,7 +127,7 @@ export function steps2To3(
     const dy = dualY[y];
     const row = matrix[y];
     for (let x = 0; x < X; ++x) {
-      if (starsX[x] === -1 && row[x] === (dualX[x] + dy || 0)) {
+      if (starsX[x] === -1 && row[x] === dualX[x] + dy) {
         starsX[x] = y;
         starsY[y] = x;
         ++stars;
@@ -150,9 +153,9 @@ export function steps2To3(
  */
 export function step4(
   unmatched: number,
-  matrix: MatrixLike<number>,
-  dualX: number[],
-  dualY: number[],
+  matrix: MatrixLike<bigint>,
+  dualX: bigint[],
+  dualY: bigint[],
   starsX: number[],
   starsY: number[]
 ): void {
@@ -164,7 +167,7 @@ export function step4(
   const Y = dualY.length;
   const coveredY = new Uint32Array(Y);
   const slack = new Uint32Array(X);
-  const slackV = new Array<number>(X);
+  const slackV = new Array<bigint>(X);
   const slackX = new Uint32Array(X);
 
   for (let rootY = 0; unmatched > 0; ++rootY) {
@@ -176,21 +179,20 @@ export function step4(
     let step = 0;
     slackX.fill(rootY);
     coveredY[0] = rootY;
-    let slackMid = initSlack(rootY, matrix, dualX, dualY, slack, slackV);
+    let zeros = initSlack(rootY, matrix, dualX, dualY, slack, slackV);
 
     // Run stage
     // eslint-disable-next-line no-constant-condition
     while (true) {
       // If no zero
-      if (step >= slackMid) {
-        // Find and zero the min
-        const min = findUncoveredMin(slackMid, slack, slackV);
-        slackMid = partition(min, slackMid, slack, slackV);
-        slackV[slack[step]] = min;
+      if (step >= zeros) {
+        // Zero the min
+        const min = findUncoveredMin(zeros, slack, slackV);
+        zeros = partition(min, zeros, slack, slackV);
       }
 
       // Prime the zero / cover the prime's column
-      const x = slack[step];
+      const x = slack[step++];
 
       // If no star in the column
       if (starsX[x] === -1) {
@@ -198,7 +200,7 @@ export function step4(
         step5(x, slackX, starsX, starsY);
 
         // Update dual variables
-        step6(step, coveredY, dualX, dualY, slack, slackV);
+        step6(step, slackV[x], coveredY, dualX, dualY, slack, slackV);
 
         // Terminate stage
         --unmatched;
@@ -207,12 +209,13 @@ export function step4(
 
       // Cover the star's row
       const sy = starsX[x];
-      coveredY[++step] = sy;
+      coveredY[step] = sy;
 
       // Update slack
-      slackMid = updateSlack(
+      zeros = updateSlack(
         sy,
-        slackMid,
+        slackV[x],
+        zeros,
         matrix,
         dualX,
         dualY,
@@ -225,79 +228,44 @@ export function step4(
 }
 
 /**
- * Augments the current matching.
+ * Adjusts dual variables to uncover more admissible edges.
  *
- * This step effectively increases the number of matches (stars)
- * by 1, bringing the algorithm closer to an optimal assignment.
- *
- * Augmentation is performed by flipping matched and unmatched edges along
- * an augmenting path, starting from an unmatched node / edge and
- * continuing until no matched edge can be found.
- *
- * @param x - The starting node's column.
- * @param primeX - An array mapping primed columns to rows.
- * @param starsX - An array mapping star columns to row. Modified in place.
- * @param starsY - An array mapping star rows to columns. Modified in place.
- */
-export function step5(
-  x: number,
-  primeX: ArrayLike<number>,
-  starsX: number[],
-  starsY: number[]
-): void {
-  do {
-    const y = primeX[x];
-    const sx = starsY[y];
-    starsX[x] = y;
-    starsY[y] = x;
-    x = sx;
-  } while (x !== -1);
-}
-
-/**
- * Adjusts dual variables and slack to uncover more admissible edges.
- *
- * @param pivot - The value to adjust by.
- * @param covV - The value indicating a row is covered.
+ * @param N - The number of adjustments to make.
+ * @param min - The value to adjust by.
  * @param coveredY - An array indicating whether a row is covered.
  * @param dualX - The dual variables associated with each column of the matrix. Modified in place.
  * @param dualY - The dual variables associated with each row of the matrix. Modified in place.
- * @param exposedX - An array indicating uncovered columns.
+ * @param slack - An array of covered column indices.
  * @param slackV - The slack values for each column. Modified in place.
  */
 export function step6(
-  max: number,
+  N: number,
+  min: bigint,
   coveredY: ArrayLike<number>,
-  dualX: number[],
-  dualY: number[],
-  slack: IndexArray,
-  slackV: number[]
+  dualX: bigint[],
+  dualY: bigint[],
+  slack: ArrayLike<number>,
+  slackV: ArrayLike<bigint>
 ): void {
-  let min = 0;
-  for (let i = 0; i <= max; ++i) {
-    min = min + slackV[slack[i]] || 0;
-  }
-
-  for (let i = 0; i <= max; ++i) {
-    let j = coveredY[i];
-    dualY[j] = dualY[j] + min || 0;
-    j = slack[i];
-    min = min - slackV[j] || 0;
-    dualX[j] = dualX[j] - min || 0;
+  let prev = 0n;
+  for (let i = 0; i < N; ++i) {
+    const x = slack[i];
+    dualY[coveredY[i]] += min - prev;
+    prev = slackV[x];
+    dualX[x] -= min - prev;
   }
 }
 
 export function partition(
-  pivot: number,
+  pivot: bigint,
   min: number,
   slack: IndexArray,
-  slackV: number[]
+  slackV: ArrayLike<bigint>
 ): number {
   const max = slack.length;
   for (let i = min; i < max; ++i) {
     const x = slack[i];
-    slackV[x] = slackV[x] - pivot || 0;
-    if (slackV[x] === 0) {
+    if (slackV[x] === pivot) {
       slack[i] = slack[min];
       slack[min++] = x;
     }
@@ -305,40 +273,13 @@ export function partition(
   return min;
 }
 
-export function findUncoveredMin(
-  mid: number,
-  slack: ArrayLike<number>,
-  slackV: ArrayLike<number>
-): number;
-export function findUncoveredMin(
-  mid: number,
-  slack: ArrayLike<number>,
-  slackV: ArrayLike<bigint>
-): bigint;
-export function findUncoveredMin<T extends number | bigint>(
-  mid: number,
-  slack: ArrayLike<number>,
-  slackV: ArrayLike<T>
-): T {
-  const X = slack.length;
-
-  let minV = slackV[slack[mid]];
-  for (let i = mid + 1; i < X; ++i) {
-    if (slackV[slack[i]] < minV) {
-      minV = slackV[slack[i]];
-    }
-  }
-
-  return minV;
-}
-
 export function initSlack(
   y: number,
-  matrix: MatrixLike<number>,
-  dualX: ArrayLike<number>,
-  dualY: ArrayLike<number>,
+  matrix: MatrixLike<bigint>,
+  dualX: ArrayLike<bigint>,
+  dualY: ArrayLike<bigint>,
   slack: IndexArray,
-  slackV: number[]
+  slackV: bigint[]
 ): number {
   const dy = dualY[y];
   const row = matrix[y];
@@ -347,8 +288,8 @@ export function initSlack(
   let mid = 0;
   for (let x = 0; x < X; ++x) {
     slack[x] = x;
-    slackV[x] = row[x] - (dualX[x] + dy || 0) || 0;
-    if (slackV[x] === 0) {
+    slackV[x] = row[x] - dualX[x] - dy;
+    if (slackV[x] === 0n) {
       slack[x] = slack[mid];
       slack[mid++] = x;
     }
@@ -357,58 +298,26 @@ export function initSlack(
   return mid;
 }
 
-/**
- * Generates a string representation of the cost matrix,
- * annotating starred (*) and primed (") elements.
- *
- * Stars (*) indicate part of a potential solution, while primes (") mark
- * elements considered for augmenting the current solution. This visualization
- * aids in understanding and debugging the matrix's state at various steps of
- * the algorithm.
- *
- * @param matrix - The cost matrix.
- * @param starsY - An array of star y coordinates to x coordinates.
- * @param primeX - (Optional) An array of prime x coordinates to y coordinates.
- *
- * @returns A string visualization of the matrix with stars and primes.
- */
-export function toString<T>(
-  matrix: Matrix<T>,
-  starsY: ArrayLike<number>,
-  primeX: ArrayLike<number>
-): string {
-  // Mark values as stars or primes
-  return _toString(matrix, (v, y, x): string => {
-    let str = `${v}`;
-    if (x == starsY[y]) {
-      str = "*" + str;
-    }
-    if (y == primeX[x]) {
-      str = '"' + str;
-    }
-    return str;
-  });
-}
-
 export function updateSlack(
   y: number,
+  minV: bigint,
   midS: number,
-  matrix: MatrixLike<number>,
-  dualX: ArrayLike<number>,
-  dualY: ArrayLike<number>,
+  matrix: MatrixLike<bigint>,
+  dualX: ArrayLike<bigint>,
+  dualY: ArrayLike<bigint>,
   slack: IndexArray,
-  slackV: number[],
+  slackV: bigint[],
   slackX: IndexArray
 ): number {
-  const dy = dualY[y];
+  const dy = dualY[y] - minV;
   const row = matrix[y];
-  const X = slack.length;
+  const X = slackX.length;
 
   for (let i = midS; i < X; ++i) {
     const x = slack[i];
-    const value = row[x] - (dualX[x] + dy || 0) || 0;
+    const value = row[x] - dualX[x] - dy;
     if (value < slackV[x]) {
-      if (value === 0) {
+      if (value === minV) {
         slack[i] = slack[midS];
         slack[midS++] = x;
       }