+++ src/index.ts
@@ -1,16 +1,16 @@
 // Types
 export type { Matrix } from "./types/matrix";
 export type { MatrixLike } from "./types/matrixLike";
-export type { Tuple } from "./types/tuple";
+export type { Pair } from "./types/pair";
 
 // Cost Matrix Helpers
 export {
-  createCostMatrix,
-  getMaxCost,
-  getMinCost,
-  invertCostMatrix,
-  negateCostMatrix,
-} from "./utils/costMatrix";
+  createMatrix,
+  getMatrixMax,
+  getMatrixMin,
+  invertMatrix,
+  negateMatrix,
+} from "./helpers";
 
 // Algorithm
 export { munkres } from "./munkres";